[v1.5.0rel] Ready!
- Added a feature to create blank installations
- Added an option to save and load configs for individual installations
- UI updates
- Code updates (optimizations, more efficient)

+++ src/cubeLauncher/options.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.IO;
+using System.Linq;
 using System.Windows.Forms;
 using System.Runtime.InteropServices;
-using System.Linq;
 
 namespace cubeLauncher
 {
@@ -19,20 +19,19 @@ namespace cubeLauncher
         // create global variables
         // path variables
         string mainDir;
-        // dynamic .cube file parser variables
+        // .cube file launcher variables
         string installName;
         string customName;
         string customVersion;
         string customWidth;
         string customHeight;
         string customArgs;
-
+        // dynamic .cube file parser variables
         string line1_name_format;
         string line2_version_format;
         string line3_width_format;
         string line4_height_format;
         string line5_arguments_format;
-        string line6_modloader_format;
 
         // form events
 
@@ -57,62 +56,97 @@ namespace cubeLauncher
             }
 
             // load configs into memory
+
+            // name
             if (File.Exists(mainDir + "\\config_name"))
             {
                 string config_name = File.ReadAllText(mainDir + "\\config_name");
                 customNameBox.Text = config_name;
             }
 
+            // version
             if (File.Exists(mainDir + "\\config_ver"))
             {
                 string config_ver = File.ReadAllText(mainDir + "\\config_ver");
                 customVersionBox.Text = config_ver;
             }
 
+            // width
             if (File.Exists(mainDir + "\\config_x"))
             {
                 string config_x = File.ReadAllText(mainDir + "\\config_x");
                 customWidthBox.Text = config_x;
             }
 
+            // height
             if (File.Exists(mainDir + "\\config_y"))
             {
                 string config_y = File.ReadAllText(mainDir + "\\config_y");
                 customHeightBox.Text = config_y;
             }
 
+            // argument
             if (File.Exists(mainDir + "\\config_args"))
             {
                 string config_args = File.ReadAllText(mainDir + "\\config_args");
                 customArgsBox.Text = config_args;
             }
 
+            // launcher path
             if (File.Exists(mainDir + "\\config_lchrpth"))
             {
                 string config_lchrpth = File.ReadAllText(mainDir + "\\config_lchrpth");
                 launcherPathLabel.Text = config_lchrpth;
             }
 
-            installName = File.ReadAllText(mainDir + "\\config_instname");
+            // configure installname
+            try
+            {
+                installName = File.ReadAllText(mainDir + "\\config_instname");
+            }
+            catch
+            {
+                // skip
+            }
+
+            // configure config label
+            if (File.Exists(mainDir + "\\config_instname") && installName != "")
+            {
+                configNameLabel.Text = "Config for \"" + installName + "\"";
+            }
+            else
+            {
+                configNameLabel.Text = "";
+            }
 
             // configure tooltips
             optionsToolTip.SetToolTip(closeButton, "Close");
-            optionsToolTip.SetToolTip(customNameBox, "Specify the name override for the launcher");
-            optionsToolTip.SetToolTip(customVersionBox, "Specify the version override for the launcher");
-            optionsToolTip.SetToolTip(customWidthBox, "Specify the resolution width override for the launcher");
-            optionsToolTip.SetToolTip(customHeightBox, "Specify the resolution height version override for the launcher");
+            optionsToolTip.SetToolTip(customNameBox, "Specify the name for the launcher");
+            optionsToolTip.SetToolTip(customVersionBox, "Specify the version for the launcher");
+            optionsToolTip.SetToolTip(customWidthBox, "Specify the resolution width for the launcher");
+            optionsToolTip.SetToolTip(customHeightBox, "Specify the resolution height for the launcher");
             optionsToolTip.SetToolTip(customArgsBox, "Specify the argument override for the launcher");
             string ovrCubeChkTT = "Override the .cube override config";
             optionsToolTip.SetToolTip(overrideCubeCheckbox, ovrCubeChkTT);
             optionsToolTip.SetToolTip(overrideCubePicture, ovrCubeChkTT);
-            optionsToolTip.SetToolTip(createConfigButton, "Create a blank .cube file and directory next to \"cubeLauncher.exe\"");
+            optionsToolTip.SetToolTip(saveConfigButton, "Save the current options to a config file specific to the selected installation");
+            optionsToolTip.SetToolTip(loadConfigButton, "Load the current config specific to the selected installation");
             optionsToolTip.SetToolTip(selectPathButton, "Select an alternative path for the Minecraft Launcher");
             optionsToolTip.SetToolTip(clearPathButton, "Reset the selected Minecraft Launcher path");
-            if (File.Exists(mainDir + "\\config_lchrpth"))
-            {
-                string lchrpth = File.ReadAllText(mainDir + "\\config_lchrpth");
-                optionsToolTip.SetToolTip(launcherPathLabel, lchrpth);
-            }
+            lchrPthTT();
+
+            // configure tooltip draw
+            optionsToolTip.OwnerDraw = true;
+            optionsToolTip.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(35)))), ((int)(((byte)(35)))));
+            optionsToolTip.ForeColor = System.Drawing.Color.FromArgb(((int)(((byte)(93)))), ((int)(((byte)(183)))), ((int)(((byte)(159)))));
+        }
+
+        // draw tooltips
+        private void optionsToolTip_Draw(object sender, DrawToolTipEventArgs e)
+        {
+            e.DrawBackground();
+            e.DrawBorder();
+            e.DrawText();
         }
 
         // buttons
@@ -150,11 +184,11 @@ namespace cubeLauncher
         }
 
         // create config button
-        private void createConfigButton_Click(object sender, EventArgs e)
+        private void saveConfigButton_Click(object sender, EventArgs e)
         {
             if (customNameBox.Text == "")
             {
-                customName = "UNNAMED";
+                customName = installName;
             }
             else
             {
@@ -200,7 +234,7 @@ namespace cubeLauncher
             try
             {
                 string cubeConfig = "# CUBELAUNCHER OVERRIDE CONFIG\nname: " + customName + "\nversion: " + customVersion + "\nwidth: " + customWidth + "\nheight: " + customHeight + "\narguments: " + customArgs + "\nmodloader: ";
-                if (installName != "")
+                if (File.Exists(mainDir + "\\config_instname") && installName != "")
                 {
                     Directory.CreateDirectory(mainDir + "\\" + installName + "\\.cube");
                     File.WriteAllText(mainDir + "\\" + installName + "\\.cube\\config.cube", cubeConfig);
@@ -209,6 +243,75 @@ namespace cubeLauncher
             catch (Exception ex)
             {
                 MessageBox.Show("Unknown Error: Unable to save \"config.cube\"!\n\nFull Error:\n" + ex);
+            }
+        }
+
+        private void loadConfigButton_Click(object sender, EventArgs e)
+        {
+            if (File.Exists(mainDir + "\\" + installName + "\\.cube\\config.cube"))
+            {
+                try
+                {
+                    string line1_name = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(1);
+                    line1_name_format = line1_name.Replace("name: ", "");
+                    customNameBox.Text = line1_name_format;
+                }
+                catch
+                {
+                    // skip
+                }
+
+                try
+                {
+                    string line2_version = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(2);
+                    line2_version_format = line2_version.Replace("version: ", "");
+                    customVersionBox.Text = line2_version_format;
+                }
+                catch
+                {
+                    // skip
+                }
+
+                try
+                {
+                    string line3_width = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(3);
+                    line3_width_format = line3_width.Replace("width: ", "");
+                    customWidthBox.Text = line3_width_format;
+                }
+                catch
+                {
+                    // skip
+                }
+
+                try
+                {
+                    string line4_height = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(4);
+                    line4_height_format = line4_height.Replace("height: ", "");
+                    customHeightBox.Text = line4_height_format;
+                }
+                catch
+                {
+                    // skip
+                }
+
+                try
+                {
+                    string line5_arguments = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(5);
+                    line5_arguments_format = line5_arguments.Replace("arguments: ", "");
+                    customArgsBox.Text = line5_arguments_format;
+                }
+                catch
+                {
+                    // skip
+                }
+            }
+            else
+            {
+                customNameBox.Text = "";
+                customVersionBox.Text = "";
+                customWidthBox.Text = "";
+                customHeightBox.Text = "";
+                customArgsBox.Text = "";
             }
         }
 
@@ -235,11 +338,7 @@ namespace cubeLauncher
                     // skip
                 }
 
-                if (File.Exists(mainDir + "\\config_lchrpth"))
-                {
-                    string lchrpth = File.ReadAllText(mainDir + "\\config_lchrpth");
-                    optionsToolTip.SetToolTip(launcherPathLabel, lchrpth);
-                }
+                lchrPthTT();
             }
         }
 
@@ -332,24 +431,6 @@ namespace cubeLauncher
             }
         }
 
-        // arguments textbox
-        private void customArgsBox_TextChanged(object sender, EventArgs e)
-        {
-            File.WriteAllText(mainDir + "\\config_args", customArgsBox.Text);
-
-            if (customArgsBox.Text == "")
-            {
-                try
-                {
-                    File.Delete(mainDir + "\\config_args");
-                }
-                catch
-                {
-                    // skip
-                }
-            }
-        }
-
         //functions
 
         // move form function
@@ -359,6 +440,16 @@ namespace cubeLauncher
             {
                 ReleaseCapture();
                 SendMessage(Handle, WM_NCLBUTTONDOWN, HT_CAPTION, 0);
+            }
+        }
+
+        // set options tooltip function
+        private void lchrPthTT()
+        {
+            if (File.Exists(mainDir + "\\config_lchrpth"))
+            {
+                string lchrpth = File.ReadAllText(mainDir + "\\config_lchrpth");
+                optionsToolTip.SetToolTip(launcherPathLabel, lchrpth);
             }
         }
 
@@ -382,59 +473,16 @@ namespace cubeLauncher
             mvFrm(e);
         }
 
-        private void loadConfigButton_Click(object sender, EventArgs e)
+        // arguments textbox
+        private void customArgsBox_TextChanged(object sender, EventArgs e)
         {
-            if (File.Exists(mainDir + "\\" + installName + "\\.cube\\config.cube"))
+            File.WriteAllText(mainDir + "\\config_args", customArgsBox.Text);
+
+            if (customArgsBox.Text == "")
             {
                 try
                 {
-                    string line1_name = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(1);
-                    line1_name_format = line1_name.Replace("name: ", "");
-                    customNameBox.Text = line1_name_format;
-                }
-                catch
-                {
-                    // skip
-                }
-
-                try
-                {
-                    string line2_version = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(2);
-                    line2_version_format = line2_version.Replace("version: ", "");
-                    customVersionBox.Text = line2_version_format;
-                }
-                catch
-                {
-                    // skip
-                }
-
-                try
-                {
-                    string line3_width = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(3);
-                    line3_width_format = line3_width.Replace("width: ", "");
-                    customWidthBox.Text = line3_width_format;
-                }
-                catch
-                {
-                    // skip
-                }
-
-                try
-                {
-                    string line4_height = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(4);
-                    line4_height_format = line4_height.Replace("height: ", "");
-                    customHeightBox.Text = line4_height_format;
-                }
-                catch
-                {
-                    // skip
-                }
-
-                try
-                {
-                    string line5_arguments = File.ReadLines(mainDir + "\\" + installName + "\\.cube\\config.cube").ElementAt(5);
-                    line5_arguments_format = line5_arguments.Replace("arguments: ", "");
-                    customArgsBox.Text = line5_arguments_format;
+                    File.Delete(mainDir + "\\config_args");
                 }
                 catch
                 {
